requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_google_genai
langchain_community
langchain-core
pyPDF
PyMuPDF
faiss-cpu
structlog
pandas
-e .